+++ packages/verse-cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @operativa/verse-cli
 
+## 0.7.2
+
+### Patch Changes
+
+- Tweak docs
+- Updated dependencies
+  - @operativa/verse@0.7.2
+  - @operativa/verse-migrations@0.7.2
+
 ## 0.7.1
 
 ### Patch Changes