+++ packages/verse-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @operativa/verse
 
+## 0.7.2
+
+### Patch Changes
+
+- Tweak docs
+
 ## 0.7.1
 
 ### Patch Changes